pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = [{ name = "Olha Devlysh", email = "odevlysh@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Add dependencies here if needed
    "pyyaml>=6.0.3",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
    "flake8>=7.3.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
]

